requirements.txt
djangorestframework==3.13.1
django-model-utils
django-filter
Django==3.2.9
Pygments==2.7.4
Markdown==3.2.1
coreapi==2.3.3
psycopg2-binary==2.9.2
django-environ==0.8.1
PyYAML==5.4
django-cockroachdb==3.2.*
channels==3.0.4
daphne==3.0.2
